fix(shader-lab): interface to class & typo
